+++ tests/expand/hash.rs
@@ -13,7 +13,7 @@ impl<T> Hash for T
 where
     T: Tuple,   // `T` must be a tuple with 1-12 components.
     T<_>: Hash, // Each component must implement `Hash`.
-    T<{ T::LEN - 1 }>: ?Sized, // The last component may be unsized.
+    T<{ T::LAST }>: ?Sized, // The last component may be unsized.
 {
     #[inline]
     fn hash<S: Hasher>(&self, state: &mut S) {